additional-studies/mini-project/ai-story-creator/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.7.1+cu128
torchvision==0.22.1+cu128
torchaudio==2.7.1+cu128
# Hugging Face 및 관련 라이브러리
transformers==4.52.4
diffusers==0.33.1
accelerate==1.7.0
bitsandbytes==0.46.0

# 기타 필수 패키지
gradio==5.34.0
sqlmodel==0.0.24
reportlab==4.4.1
pillow==11.2.1

# character_lora_fine_tuning 사용 시 필요
peft==0.15.2
